requirements.txt
requests
python-dotenv
fastapi
uvicorn
mysql-connector-python
# facultatif pour bonus
bandit -r
audit



